part2_job_source/requirements.txt
apify-client>=1.7
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv
# Optional: AI browser agent (Strategy 4)
browser-use>=0.1
langchain-openai>=0.2
